Delete Unnecessary code
