+++ lib/presentation/pages/news/secret.dart
@@ -1,2 +1,2 @@
 String apiKey = "3710bda2d0e34f75b30c6de80e8610ea";
-//апи ключ для проверки работоспособности
+//апи ключ для проверки работоспособности новостей
