+++ lib/presentation/pages/news/widget_CategoryCard.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+
+class CategoryCard extends StatelessWidget {
+  final String imageAssetUrl, categoryName;
+
+  const CategoryCard(
+      {Key? key, required this.imageAssetUrl, required this.categoryName})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      child: Container(
+        margin: EdgeInsets.only(right: 14, top: 8),
+        child: Stack(
+          children: <Widget>[
+            Container(
+              alignment: Alignment.center,
+              height: 35,
+              width: 100,
+              decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(10), color: Colors.blue),
+              child: Text(
+                categoryName,
+                textAlign: TextAlign.center,
+                style: TextStyle(
+                    color: Colors.white,
+                    fontSize: 14,
+                    fontWeight: FontWeight.w500),
+              ),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
